requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
idna==3.3
    # via requests
numpy==1.22.3
    # via pandas
pandas==1.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via pandas
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
urllib3==1.26.9
    # via requests
